SIH_FINAL_PROJECT/Code_on_track/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
minio==7.2.0
qrcode[pil]==7.4.2
python-magic==0.4.27
aiofiles==23.2.1
pydantic==2.5.1
pydantic-settings==2.0.3
python-dateutil==2.8.2
